pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icloud-mcp"
version = "0.1.0"
description = "MCP server for iCloud Calendar, Mail, and Reminders"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "caldav>=1.3.0",
    "icalendar>=5.0.0",
    "pyobjc-framework-EventKit>=10.0",
    "keyring>=25.0.0",
    "python-dateutil>=2.9.0",
]

[project.scripts]
icloud-mcp = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
